requirements.txt
fastapi
uvicorn
sqlalchemy
mysql-connector-python>=8.0.0
passlib[bcrypt]
python-jose
pydantic
python-dotenv
email_validator
streamlit
requests
pandas
